[CIRC-562] Rename OverdueFineCalculatorService to OverdueFineService
